+++ CHANGELOG.md
@@ -4,6 +4,7 @@
  - [OTHER] Achievement description tooltip now also shows its state
  - [OTHER] Achievements window has updated description
  - [FEATURE] Game now auto mutes sound when not visible
+ - [FEATURE] Mouse cursor now hides when playing the game and reappears when moved
 
 ## 1.6.12 - 2025-05-19
  - [BUGFIX] Minimap will correctly render if you first restore to a room and then use restore function without leaving the room